requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
frozenlist==1.3.3
idna==3.4
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.4
packaging==23.0
pip==23.0.1
prompt-toolkit==3.0.38
PyYAML==6.0
requests==2.28.2
setuptools==67.6.0
tqdm==4.65.0
TTS==0.11.1
typing_extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.15
wcwidth==0.2.6
wheel==0.40.0
yamldataclassconfig==1.5.0
yarl==1.8.2
